requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cookiebanner==0.2.8
django-solo==2.2.0
pillow==10.3.0
python-dotenv==1.0.1
sqlparse==0.5.0
